Fixed warnings
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@2430 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/client/CClient_Draw.cpp
@@ -1371,7 +1371,7 @@ void CClient::DrawGameMenu(SDL_Surface * bmpDest)
 						info->addItem(3, "Received: " + itoa(w->getClient()->getChannel()->getIncoming()/1024) + " Kb");
 						info->addItem(4, "Sent: " + itoa(w->getClient()->getChannel()->getOutgoing()/1024) + " Kb");
 						info->addItem(5, "Connected for " + 
-							itoa( (tLX->fCurTime - w->getClient()->getConnectTime()) / 60.0f) + " minutes");
+							itoa( (int)((tLX->fCurTime - w->getClient()->getConnectTime()) / 60.0f)) + " minutes");
 						info->addItem(6, "Total " + 
 							itoa( w->getTotalKills() ) + " kills / " +
 							itoa( w->getTotalDeaths() ) + " deaths / " +
